+++ tests/common.rs
@@ -1,5 +1,9 @@
 #![allow(dead_code)]
-use std::{fs, path::Path};
+
+use std::{
+    fs, io,
+    path::{Path, PathBuf},
+};
 
 use dom_query::{Document, Matcher};
 use dom_smoothie::{CandidateSelectMode, Config, Readability, TextMode};
@@ -7,6 +11,36 @@ use dom_smoothie::{CandidateSelectMode, Config, Readability, TextMode};
 use once_cell::sync::Lazy;
 pub(crate) static R_MATCHER: Lazy<Matcher> =
     Lazy::new(|| Matcher::new("#readability-page-1").unwrap());
+
+macro_rules! check {
+    ($field:expr, $left:expr, $right:expr, $path:expr) => {
+        assert_eq!(
+            $left,
+            $right,
+            "Mismatch in field '{}' (test: {})",
+            $field,
+            $path.display()
+        );
+    };
+}
+
+#[allow(unused_macros)]
+macro_rules! test_data {
+    ($test_path:expr, $expected_file:expr) => {
+        TestData::from_path($test_path, None, $expected_file).unwrap()
+    };
+}
+
+#[allow(unused_macros)]
+macro_rules! include_test_data {
+    ($test_path:expr, $source_file:expr, $expected_file:expr) => {
+        TestData::new(
+            $test_path,
+            include_str!($source_file).to_string(),
+            include_str!($expected_file).to_string(),
+        )
+    };
+}
 
 #[derive(Debug, serde::Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -22,13 +56,44 @@ struct ExpectedMetadata {
     image: Option<String>,
 }
 
-pub(crate) fn test_alt_text<P>(test_path: P, text_mode: TextMode, expected_filename: &str)
-where
-    P: AsRef<Path>,
-{
-    let base_path = test_path.as_ref();
-    let source_path = base_path.join("source.html");
-    let expected_path = base_path.join(expected_filename);
+pub struct TestData {
+    path: PathBuf,
+    source_contents: String,
+    expected_contents: String,
+}
+
+impl TestData {
+    /// new creates a new instance of `TestData`.
+    pub fn new<P>(test_path: P, source_contents: String, expected_contents: String) -> Self
+    where
+        P: AsRef<Path>,
+    {
+        Self {
+            path: test_path.as_ref().to_path_buf(),
+            source_contents,
+            expected_contents,
+        }
+    }
+
+    pub fn from_path<P>(
+        test_path: P,
+        source_file: Option<&str>,
+        expected_file: &str,
+    ) -> io::Result<Self>
+    where
+        P: AsRef<Path>,
+    {
+        let base_path = test_path.as_ref();
+        let source_file = source_file.unwrap_or("source.html");
+        let source_path = base_path.join(source_file);
+        let source_contents = fs::read_to_string(source_path)?;
+        let expected_path = base_path.join(expected_file);
+        let expected_contents = fs::read_to_string(expected_path)?;
+        Ok(Self::new(test_path, source_contents, expected_contents))
+    }
+}
+
+pub(crate) fn test_alt_text(data: TestData, text_mode: TextMode) {
     // for more options check the documentation
     let cfg = Config {
         candidate_select_mode: CandidateSelectMode::DomSmoothie,
@@ -36,37 +101,27 @@ where
         ..Default::default()
     };
 
-    let source_contents = fs::read_to_string(source_path).unwrap();
-    let mut readability = Readability::new(source_contents, None, Some(cfg)).unwrap();
+    let mut readability = Readability::new(data.source_contents, None, Some(cfg)).unwrap();
 
     let article = readability.parse().unwrap();
-    let expected_contents = fs::read_to_string(expected_path).unwrap();
     let article_text = article.text_content.as_ref();
-    assert_eq!(article_text, expected_contents.trim())
+    let expected = data.expected_contents.trim();
+    check!("text_content",article_text, expected, &data.path);
 }
 
-pub(crate) fn test_readability<P>(test_path: P)
-where
-    P: AsRef<Path>,
-{
+pub(crate) fn test_readability(data: TestData) {
     let doc_url = Some("http://fakehost/test/");
-    let base_path = test_path.as_ref();
-    let source_path = base_path.join("source.html");
-    let expected_path = base_path.join("expected.html");
-
-    let source_contents = fs::read_to_string(source_path).unwrap();
     let cfg = dom_smoothie::Config {
         classes_to_preserve: vec!["caption".into()],
         ..Default::default()
     };
-    let mut r = Readability::new(source_contents, doc_url, Some(cfg)).unwrap();
+    let mut r = Readability::new(data.source_contents, doc_url, Some(cfg)).unwrap();
     let article = r.parse().unwrap();
 
     let contents = article.content;
     let article_doc = Document::from(contents);
 
-    let expected_contents = fs::read_to_string(expected_path).unwrap();
-    let expected_doc = Document::from(expected_contents);
+    let expected_doc = Document::from(data.expected_contents);
 
     let a_html = article_doc
         .select_single_matcher(&R_MATCHER)
@@ -82,77 +137,44 @@ where
         .collect::<Vec<_>>()
         .join("");
 
-    assert_eq!(
-        a_html,
-        e_html,
-        "parsed contents for test {} do not match with expected content",
-        test_path.as_ref().display()
-    );
+    check!("content", a_html, e_html, data.path);
 }
 
-pub fn test_metadata<P>(test_path: P, host: Option<&str>)
-where
-    P: AsRef<Path>,
-{
-    let base_path = test_path.as_ref();
-    let source_path = base_path.join("source.html");
-
-    let source_contents = fs::read_to_string(source_path).unwrap();
+pub fn test_metadata(data: TestData, host: Option<&str>) {
     let cfg = dom_smoothie::Config {
         classes_to_preserve: vec!["caption".into()],
         ..Default::default()
     };
-    let mut r = Readability::new(source_contents, host, Some(cfg)).unwrap();
+    let mut r = Readability::new(data.source_contents, host, Some(cfg)).unwrap();
 
     let readable = r.is_probably_readable();
-
     let article = r.parse().unwrap();
 
-    let expected_metadata_path = base_path.join("expected-metadata.json");
-    let meta_contents = fs::read_to_string(expected_metadata_path).unwrap();
-    let expected: ExpectedMetadata = serde_json::from_str(&meta_contents).unwrap();
+    let exp: ExpectedMetadata = serde_json::from_str(&data.expected_contents).unwrap();
 
-    assert_eq!(
-        readable, expected.readerable,
-        "readerable does not match expected"
+    check!("readable", &readable, &exp.readerable, data.path);
+    check!("title", &article.title, &exp.title, data.path);
+    check!("byline", &article.byline, &exp.byline, data.path);
+    check!("excerpt", &article.excerpt, &exp.excerpt, data.path);
+    check!("site_name", &article.site_name, &exp.site_name, data.path);
+    check!(
+        "published_time",
+        &article.published_time,
+        &exp.published_time,
+        data.path
     );
-
-    assert_eq!(
-        article.title, expected.title,
-        "title does not match expected"
-    );
-    assert_eq!(
-        article.byline, expected.byline,
-        "byline does not match expected"
-    );
-    assert_eq!(
-        article.excerpt, expected.excerpt,
-        "excerpt does not match expected"
-    );
-    assert_eq!(
-        article.site_name, expected.site_name,
-        "site_name does not match expected"
-    );
-    assert_eq!(
-        article.published_time, expected.published_time,
-        "published_time does not match expected"
-    );
-    assert_eq!(article.lang, expected.lang, "lang does not match expected");
-    assert_eq!(article.dir, expected.dir, "dirs does not match expected");
-    assert_eq!(
-        article.image, expected.image,
-        "image does not match expected"
-    );
+    check!("lang", &article.lang, &exp.lang, data.path);
+    check!("dir", &article.dir, &exp.dir, data.path);
+    check!("image", &article.image, &exp.image, data.path);
 }
 
 pub fn test_favicon<P>(test_path: P, host: Option<&str>, expected: Option<&str>)
 where
     P: AsRef<Path>,
 {
-    let base_path = test_path.as_ref();
-    let source_path = base_path.join("source.html");
-
+    let source_path = test_path.as_ref().join("source.html");
     let source_contents = fs::read_to_string(source_path).unwrap();
+
     let cfg = dom_smoothie::Config {
         classes_to_preserve: vec!["caption".into()],
         ..Default::default()
@@ -161,9 +183,10 @@ where
 
     let metadata = r.get_article_metadata(None);
 
-    assert_eq!(
-        metadata.favicon,
-        expected.map(|s| s.to_string()),
-        "favicon does not match expected"
+    check!(
+        "favicon",
+        &metadata.favicon,
+        &expected.map(|s| s.to_string()),
+        &test_path.as_ref()
     );
 }